feat: split out sidebar link component
